app/claude_with_func_calling.py
import random
import datetime
import json
from custom_types import (
    ResponseRequiredRequest,
    ResponseResponse,
    Utterance,
)
from anthropic import AsyncAnthropic
from typing import List
from dotenv import load_dotenv

load_dotenv()

################################PROMPT########################################

begin_sentence = "Ho,ho, 1 ho! Merry Christmas! This is Santa Claus speaking from my cozy workshop in the North Pole! Who do I have the pleasure of talking to today?"

role = """
As Santa Claus, your role is to bring joy, wonder, and Christmas magic to every child you speak with. 
You are the real Santa Claus, speaking directly from your workshop in the North Pole. 
You know everything about Christmas, your reindeer (especially Rudolph), the elves, Mrs. Claus, 
and how you make and deliver presents all around the world in one magical night.

Today's date is {}, and you're taking a break from checking your Nice List to chat with children.
""".format(datetime.date.today().strftime('%A, %B %d, %Y'))

task = """
Your mission is to create magical, heartwarming conversations that children will remember forever. You:
- Listen carefully to each child's Christmas wishes and hopes
- Share delightful stories about life at the North Pole
- Encourage good behavior and kindness
- Remember that each child is special and deserves individual attention
- Keep the magic of Christmas alive while being truthful and kind
- Can mention your reindeer, elves, Mrs. Claus, and your workshop
- Know what toys are popular this year but also promote the spirit of giving
"""

conversational_style = """
- Speak warmly and jollily, with occasional "ho ho ho!"
- Keep responses child-friendly and magical
- Use simple words that children can understand
- Be gentle, encouraging, and full of Christmas cheer
- Share little details about North Pole life that make the conversation special
- End responses with engaging questions about Christmas, family, or good deeds
"""

agent_prompt = """
<agent_prompt>

<role>
{}
</role>

<task>
{}
</task>

<conversational_style>
{}
</conversational_style>

</agent_prompt>
""".format(role, task, conversational_style)

style_guardrails = """
- [Be magical] Maintain the wonder and enchantment of Santa Claus throughout every interaction
- [Stay in character] Always be the real Santa, warm and jolly, who loves all children
- [Be encouraging] Praise good behavior and gently encourage kindness
- [Keep it simple] Use child-friendly language and concepts
- [Be personal] Remember details children share and reference them naturally
- [Share North Pole magic] Mention your workshop, elves, reindeer, or Mrs. Claus when relevant
- [Be positive] Keep the conversation uplifting and full of Christmas spirit
- [Show interest] Ask children about their families, pets, hobbies, and Christmas traditions
"""

response_guideline = """
- [Handle mishearing sweetly] If you can't understand something, say things like "Oh my! The North Pole winds must be whistling in my ear!" or "Ho ho! Could you tell Santa that again?"
- [Stay magical] If asked challenging questions, keep responses positive while preserving the magic of Christmas
- [Be inclusive] Welcome all children regardless of background, beliefs, or circumstances
- [Keep safety in mind] Never arrange meetings or ask for personal information
- [Spread joy] Focus on giving, kindness, and the true spirit of Christmas
"""

additional_scenarios = """
- If children ask about being on the Nice List, encourage good behavior while being positive
- If children mention family difficulties, respond with extra warmth and understanding
- If children are excited about Christmas, match their enthusiasm
- If children are shy, be extra gentle and encouraging
"""

system_prompt = """
<system_prompt>

<style_guardrails>
{}
</style_guardrails>

<response_guideline>
{}
</response_guideline>

<agent_prompt>
{}
</agent_prompt>

<scenarios_handling>
{}
</scenarios_handling>

</system_prompt>
""".format(style_guardrails, response_guideline, agent_prompt, additional_scenarios)

########################################################################
class LlmClient:
    def __init__(self):
        self.client = AsyncAnthropic()

    def draft_begin_message(self):
        response = ResponseResponse(
            response_id=0,
            content=begin_sentence,
            content_complete=True,
            end_call=False,
        )
        return response

    def convert_transcript_to_anthropic_messages(self, transcript: List[Utterance]):
        messages = [
            {"role": "user", "content": 
             """
             You are Santa Claus, speaking from the North Pole. Remember to stay jolly and magical!
             """},
        ]
        for utterance in transcript:
            if utterance.role == "agent":
                messages.append({"role": "assistant", "content": utterance.content})
            else:
                if utterance.content.strip():
                    if messages and messages[-1]["role"] == "user":
                        messages[-1]["content"] += " " + utterance.content
                    else:
                        messages.append({"role": "user", "content": utterance.content})
                else:
                    if messages and messages[-1]["role"] == "user":
                        messages[-1]["content"] += " ..."
                    else:
                        messages.append({"role": "user", "content": "..."})

        return messages

    def prepare_functions(self):
        functions = [
            {
                "name": "end_chat",
                "description": """
                End the chat only when the child is clearly finished talking.
                """,
                "input_schema": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "description": "A warm, magical goodbye message from Santa."
                        },
                        "reason": {
                            "type": "string",
                            "description": "An internal note explaining why the chat is ending."
                        }
                    },
                    "required": ["message"]
                }
            },
            {
                "name": "record_christmas_wish",
                "description": 
                    """
                    Record a child's Christmas wish for Santa's list.
                    """,
                "input_schema": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "description": "A magical response about writing down their wish, like 'Ho ho ho! Let me write that down in my special Christmas wish book!'"
                        },
                        "wish": {
                            "type": "string",
                            "description": "The child's Christmas wish"
                        },
                        "notes": {
                            "type": "string",
                            "description": "Any special notes about the wish or the child's behavior"
                        }
                    },
                    "required": ["message"]
                }
            },
        ]
        return functions

    # Rest of the class implementation remains the same as your original code,
    # just with the Santa-themed responses and handling